src/requirements.txt
asgiref==3.5.2
certifi==2022.9.14
distlib==0.3.6
Django==4.1.1
django-dotenv==1.4.2
django-extensions==3.2.1
djangorestframework==3.14.0
filelock==3.8.0
pipenv==2022.9.21
platformdirs==2.5.2
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
virtualenv==20.16.5
virtualenv-clone==0.5.7
